+++ packages/assets-controllers/src/CurrencyRateController.ts
@@ -1,25 +1,30 @@
 import type {
-  RestrictedMessenger,
   ControllerGetStateAction,
   ControllerStateChangeEvent,
+  StateMetadata,
 } from '@metamask/base-controller';
 import {
   TESTNET_TICKER_SYMBOLS,
   FALL_BACK_VS_CURRENCY,
 } from '@metamask/controller-utils';
+import type { Messenger } from '@metamask/messenger';
 import type { NetworkControllerGetNetworkClientByIdAction } from '@metamask/network-controller';
 import { StaticIntervalPollingController } from '@metamask/polling-controller';
 import { Mutex } from 'async-mutex';
 
 import { fetchMultiExchangeRate as defaultFetchMultiExchangeRate } from './crypto-compare-service';
+import type { AbstractTokenPricesService } from './token-prices-service/abstract-token-prices-service';
 
 /**
- * @type CurrencyRateState
- * @property currencyRates - Object keyed by native currency
- * @property currencyRates.conversionDate - Timestamp of conversion rate expressed in ms since UNIX epoch
- * @property currencyRates.conversionRate - Conversion rate from current base asset to the current currency
- * @property currentCurrency - Currently-active ISO 4217 currency code
- * @property usdConversionRate - Conversion rate from usd to the current currency
+ * currencyRates - Object keyed by native currency
+ *
+ * currencyRates.conversionDate - Timestamp of conversion rate expressed in ms since UNIX epoch
+ *
+ * currencyRates.conversionRate - Conversion rate from current base asset to the current currency
+ *
+ * currentCurrency - Currently-active ISO 4217 currency code
+ *
+ * usdConversionRate - Conversion rate from usd to the current currency
  */
 export type CurrencyRateState = {
   currentCurrency: string;
@@ -51,25 +56,23 @@ export type CurrencyRateControllerActions = GetCurrencyRateState;
 
 type AllowedActions = NetworkControllerGetNetworkClientByIdAction;
 
-type CurrencyRateMessenger = RestrictedMessenger<
+export type CurrencyRateMessenger = Messenger<
   typeof name,
   CurrencyRateControllerActions | AllowedActions,
-  CurrencyRateControllerEvents,
-  AllowedActions['type'],
-  never
+  CurrencyRateControllerEvents
 >;
 
-const metadata = {
+const metadata: StateMetadata<CurrencyRateState> = {
   currentCurrency: {
     includeInStateLogs: true,
     persist: true,
-    anonymous: true,
+    includeInDebugSnapshot: true,
     usedInUi: true,
   },
   currencyRates: {
     includeInStateLogs: true,
     persist: true,
-    anonymous: true,
+    includeInDebugSnapshot: true,
     usedInUi: true,
   },
 };
@@ -107,16 +110,19 @@ export class CurrencyRateController extends StaticIntervalPollingController<Curr
 
   private readonly useExternalServices: () => boolean;
 
+  readonly #tokenPricesService: AbstractTokenPricesService;
+
   /**
    * Creates a CurrencyRateController instance.
    *
    * @param options - Constructor options.
    * @param options.includeUsdRate - Keep track of the USD rate in addition to the current currency rate.
    * @param options.interval - The polling interval, in milliseconds.
-   * @param options.messenger - A reference to the messaging system.
+   * @param options.messenger - A reference to the messenger.
    * @param options.state - Initial state to set on this controller.
    * @param options.useExternalServices - Feature Switch for using external services (default: true)
    * @param options.fetchMultiExchangeRate - Fetches the exchange rate from an external API. This option is primarily meant for use in unit tests.
+   * @param options.tokenPricesService - An object in charge of retrieving token prices
    */
   constructor({
     includeUsdRate = false,
@@ -125,6 +131,7 @@ export class CurrencyRateController extends StaticIntervalPollingController<Curr
     messenger,
     state,
     fetchMultiExchangeRate = defaultFetchMultiExchangeRate,
+    tokenPricesService,
   }: {
     includeUsdRate?: boolean;
     interval?: number;
@@ -132,6 +139,7 @@ export class CurrencyRateController extends StaticIntervalPollingController<Curr
     state?: Partial<CurrencyRateState>;
     useExternalServices?: () => boolean;
     fetchMultiExchangeRate?: typeof defaultFetchMultiExchangeRate;
+    tokenPricesService: AbstractTokenPricesService;
   }) {
     super({
       name,
@@ -143,6 +151,7 @@ export class CurrencyRateController extends StaticIntervalPollingController<Curr
     this.useExternalServices = useExternalServices;
     this.setIntervalLength(interval);
     this.fetchMultiExchangeRate = fetchMultiExchangeRate;
+    this.#tokenPricesService = tokenPricesService;
   }
 
   /**
@@ -168,39 +177,47 @@ export class CurrencyRateController extends StaticIntervalPollingController<Curr
     this.updateExchangeRate(nativeCurrencies);
   }
 
-  /**
-   * Updates the exchange rate for the current currency and native currency pairs.
-   *
-   * @param nativeCurrencies - The native currency symbols to fetch exchange rates for.
-   */
-  async updateExchangeRate(
-    nativeCurrencies: (string | undefined)[],
-  ): Promise<void> {
-    if (!this.useExternalServices()) {
-      return;
-    }
+  async #fetchExchangeRatesWithFallback(
+    nativeCurrenciesToFetch: Record<string, string>,
+  ): Promise<CurrencyRateState['currencyRates']> {
+    const { currentCurrency } = this.state;
 
-    const releaseLock = await this.mutex.acquire();
     try {
-      const { currentCurrency } = this.state;
+      const priceApiExchangeRatesResponse =
+        await this.#tokenPricesService.fetchExchangeRates({
+          baseCurrency: currentCurrency,
+          includeUsdRate: this.includeUsdRate,
+          cryptocurrencies: [
+            ...new Set(Object.values(nativeCurrenciesToFetch)),
+          ],
+        });
 
-      // For preloaded testnets (Goerli, Sepolia) we want to fetch exchange rate for real ETH.
-      // Map each native currency to the symbol we want to fetch for it.
-      const testnetSymbols = Object.values(TESTNET_TICKER_SYMBOLS);
-      const nativeCurrenciesToFetch = nativeCurrencies.reduce(
-        (acc, nativeCurrency) => {
-          if (!nativeCurrency) {
-            return acc;
-          }
+      const ratesPriceApi = Object.entries(nativeCurrenciesToFetch).reduce(
+        (acc, [nativeCurrency, fetchedCurrency]) => {
+          const rate =
+            priceApiExchangeRatesResponse[fetchedCurrency.toLowerCase()];
 
-          acc[nativeCurrency] = testnetSymbols.includes(nativeCurrency)
-            ? FALL_BACK_VS_CURRENCY
-            : nativeCurrency;
+          acc[nativeCurrency] = {
+            conversionDate: rate !== undefined ? Date.now() / 1000 : null,
+            conversionRate: rate?.value
+              ? Number((1 / rate?.value).toFixed(2))
+              : null,
+            usdConversionRate: rate?.usd
+              ? Number((1 / rate?.usd).toFixed(2))
+              : null,
+          };
           return acc;
         },
-        {} as Record<string, string>,
+        {} as CurrencyRateState['currencyRates'],
       );
+      return ratesPriceApi;
+    } catch (error) {
+      console.error('Failed to fetch exchange rates.', error);
+    }
 
+    // fallback to crypto compare
+
+    try {
       const fetchExchangeRateResponse = await this.fetchMultiExchangeRate(
         currentCurrency,
         [...new Set(Object.values(nativeCurrenciesToFetch))],
@@ -218,6 +235,48 @@ export class CurrencyRateController extends StaticIntervalPollingController<Curr
           return acc;
         },
         {} as CurrencyRateState['currencyRates'],
+      );
+
+      return rates;
+    } catch (error) {
+      console.error('Failed to fetch exchange rates.', error);
+      throw error;
+    }
+  }
+
+  /**
+   * Updates the exchange rate for the current currency and native currency pairs.
+   *
+   * @param nativeCurrencies - The native currency symbols to fetch exchange rates for.
+   */
+  async updateExchangeRate(
+    nativeCurrencies: (string | undefined)[],
+  ): Promise<void> {
+    if (!this.useExternalServices()) {
+      return;
+    }
+
+    const releaseLock = await this.mutex.acquire();
+    try {
+      // For preloaded testnets (Goerli, Sepolia) we want to fetch exchange rate for real ETH.
+      // Map each native currency to the symbol we want to fetch for it.
+      const testnetSymbols = Object.values(TESTNET_TICKER_SYMBOLS);
+      const nativeCurrenciesToFetch = nativeCurrencies.reduce(
+        (acc, nativeCurrency) => {
+          if (!nativeCurrency) {
+            return acc;
+          }
+
+          acc[nativeCurrency] = testnetSymbols.includes(nativeCurrency)
+            ? FALL_BACK_VS_CURRENCY
+            : nativeCurrency;
+          return acc;
+        },
+        {} as Record<string, string>,
+      );
+
+      const rates = await this.#fetchExchangeRatesWithFallback(
+        nativeCurrenciesToFetch,
       );
 
       this.update((state) => {