+++ packages/bridge-controller/src/types.ts
@@ -5,12 +5,12 @@ import type {
   TokenRatesControllerGetStateAction,
 } from '@metamask/assets-controllers';
 import type {
+  ControllerGetStateAction,
   ControllerStateChangeEvent,
-  RestrictedMessenger,
 } from '@metamask/base-controller';
+import type { Messenger } from '@metamask/messenger';
 import type {
   NetworkControllerFindNetworkClientIdByChainIdAction,
-  NetworkControllerGetStateAction,
   NetworkControllerGetNetworkClientByIdAction,
 } from '@metamask/network-controller';
 import type { RemoteFeatureFlagControllerGetStateAction } from '@metamask/remote-feature-flag-controller';
@@ -216,7 +216,10 @@ export type QuoteRequest<
    * Whether to request quotes that use EIP-7702 delegated gasless execution
    */
   gasIncluded7702: boolean;
-  noFee?: boolean;
+  /**
+   * The fee that will be charged by MetaMask
+   */
+  fee?: number;
 };
 
 export enum StatusTypes {
@@ -344,8 +347,19 @@ export type BridgeControllerAction<
   handler: BridgeController[FunctionName];
 };
 
+export type BridgeControllerGetStateAction = ControllerGetStateAction<
+  typeof BRIDGE_CONTROLLER_NAME,
+  BridgeControllerState
+>;
+
+export type BridgeControllerStateChangeEvent = ControllerStateChangeEvent<
+  typeof BRIDGE_CONTROLLER_NAME,
+  BridgeControllerState
+>;
+
 // Maps to BridgeController function names
 export type BridgeControllerActions =
+  | BridgeControllerGetStateAction
   | BridgeControllerAction<BridgeBackgroundAction.SET_CHAIN_INTERVAL_LENGTH>
   | BridgeControllerAction<BridgeBackgroundAction.RESET_STATE>
   | BridgeControllerAction<BridgeBackgroundAction.GET_BRIDGE_ERC20_ALLOWANCE>
@@ -354,10 +368,7 @@ export type BridgeControllerActions =
   | BridgeControllerAction<BridgeBackgroundAction.FETCH_QUOTES>
   | BridgeControllerAction<BridgeUserAction.UPDATE_QUOTE_PARAMS>;
 
-export type BridgeControllerEvents = ControllerStateChangeEvent<
-  typeof BRIDGE_CONTROLLER_NAME,
-  BridgeControllerState
->;
+export type BridgeControllerEvents = BridgeControllerStateChangeEvent;
 
 export type AllowedActions =
   | AccountsControllerGetAccountByAddressAction
@@ -366,7 +377,6 @@ export type AllowedActions =
   | MultichainAssetsRatesControllerGetStateAction
   | HandleSnapRequest
   | NetworkControllerFindNetworkClientIdByChainIdAction
-  | NetworkControllerGetStateAction
   | NetworkControllerGetNetworkClientByIdAction
   | RemoteFeatureFlagControllerGetStateAction;
 export type AllowedEvents = never;
@@ -374,10 +384,8 @@ export type AllowedEvents = never;
 /**
  * The messenger for the BridgeController.
  */
-export type BridgeControllerMessenger = RestrictedMessenger<
+export type BridgeControllerMessenger = Messenger<
   typeof BRIDGE_CONTROLLER_NAME,
   BridgeControllerActions | AllowedActions,
-  BridgeControllerEvents | AllowedEvents,
-  AllowedActions['type'],
-  AllowedEvents['type']
+  BridgeControllerEvents | AllowedEvents
 >;