+++ packages/bridge-status-controller/src/bridge-status-controller.ts
@@ -82,7 +82,7 @@ const metadata: StateMetadata<BridgeStatusControllerState> = {
   txHistory: {
     includeInStateLogs: true,
     persist: true,
-    anonymous: false,
+    includeInDebugSnapshot: false,
     usedInUi: true,
   },
 };
@@ -168,27 +168,27 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
     this.#trace = traceFn ?? (((_request, fn) => fn?.()) as TraceCallback);
 
     // Register action handlers
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${BRIDGE_STATUS_CONTROLLER_NAME}:startPollingForBridgeTxStatus`,
       this.startPollingForBridgeTxStatus.bind(this),
     );
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${BRIDGE_STATUS_CONTROLLER_NAME}:wipeBridgeStatus`,
       this.wipeBridgeStatus.bind(this),
     );
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${BRIDGE_STATUS_CONTROLLER_NAME}:resetState`,
       this.resetState.bind(this),
     );
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${BRIDGE_STATUS_CONTROLLER_NAME}:submitTx`,
       this.submitTx.bind(this),
     );
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${BRIDGE_STATUS_CONTROLLER_NAME}:restartPollingForFailedAttempts`,
       this.restartPollingForFailedAttempts.bind(this),
     );
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${BRIDGE_STATUS_CONTROLLER_NAME}:getBridgeHistoryItemByTxMetaId`,
       this.getBridgeHistoryItemByTxMetaId.bind(this),
     );
@@ -196,7 +196,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
     // Set interval
     this.setIntervalLength(REFRESH_INTERVAL_MS);
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'TransactionController:transactionFailed',
       ({ transactionMeta }) => {
         const { type, status, id } = transactionMeta;
@@ -228,7 +228,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
       },
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'TransactionController:transactionConfirmed',
       (transactionMeta) => {
         const { type, id, chainId } = transactionMeta;
@@ -284,10 +284,10 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
         state.txHistory = DEFAULT_BRIDGE_STATUS_CONTROLLER_STATE.txHistory;
       });
     } else {
-      const { selectedNetworkClientId } = this.messagingSystem.call(
+      const { selectedNetworkClientId } = this.messenger.call(
         'NetworkController:getState',
       );
-      const selectedNetworkClient = this.messagingSystem.call(
+      const selectedNetworkClient = this.messenger.call(
         'NetworkController:getNetworkClientById',
         selectedNetworkClientId,
       );
@@ -524,7 +524,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
   };
 
   #getMultichainSelectedAccount(accountAddress: string) {
-    return this.messagingSystem.call(
+    return this.messenger.call(
       'AccountsController:getAccountByAddress',
       accountAddress,
     );
@@ -651,7 +651,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
             UnifiedSwapBridgeEventName.Completed,
             bridgeTxMetaId,
           );
-          this.messagingSystem.publish(
+          this.messenger.publish(
             'BridgeStatusController:destinationTransactionCompleted',
             historyItem.quote.destAsset.assetId,
           );
@@ -680,7 +680,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
     }
 
     // Look up in TransactionController if txMeta has been updated with the srcTxHash
-    const txControllerState = this.messagingSystem.call(
+    const txControllerState = this.messenger.call(
       'TransactionController:getState',
     );
     const txMeta = txControllerState.transactions.find(
@@ -771,7 +771,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
     }
 
     const request = getClientRequest(quoteResponse, selectedAccount);
-    const requestResponse = (await this.messagingSystem.call(
+    const requestResponse = (await this.messenger.call(
       'SnapController:handleRequest',
       request,
     )) as
@@ -797,12 +797,9 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
     >['result'],
   ): Promise<TransactionMeta> => {
     const transactionHash = await hashPromise;
-    const finalTransactionMeta: TransactionMeta | undefined =
-      this.messagingSystem
-        .call('TransactionController:getState')
-        .transactions.find(
-          (tx: TransactionMeta) => tx.hash === transactionHash,
-        );
+    const finalTransactionMeta: TransactionMeta | undefined = this.messenger
+      .call('TransactionController:getState')
+      .transactions.find((tx: TransactionMeta) => tx.hash === transactionHash);
     if (!finalTransactionMeta) {
       throw new Error(
         'Failed to submit cross-chain swap tx: txMeta for txHash was not found',
@@ -871,7 +868,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
   }): Promise<TransactionMeta> => {
     const actionId = generateActionId().toString();
 
-    const selectedAccount = this.messagingSystem.call(
+    const selectedAccount = this.messenger.call(
       'AccountsController:getAccountByAddress',
       trade.from,
     );
@@ -881,7 +878,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
       );
     }
     const hexChainId = formatChainIdToHex(trade.chainId);
-    const networkClientId = this.messagingSystem.call(
+    const networkClientId = this.messenger.call(
       'NetworkController:findNetworkClientIdByChainId',
       hexChainId,
     );
@@ -922,7 +919,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
     quoteResponse: QuoteResponse & Partial<QuoteMetadata>,
   ) => {
     const resetApproval = await getUSDTAllowanceResetTx(
-      this.messagingSystem,
+      this.messenger,
       quoteResponse,
     );
     if (resetApproval) {
@@ -938,7 +935,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
     networkClientId: string,
     chainId: Hex,
   ) => {
-    const { gasFeeEstimates } = this.messagingSystem.call(
+    const { gasFeeEstimates } = this.messenger.call(
       'GasFeeController:getState',
     );
     const { estimates: txGasFeeEstimates } = await this.#estimateGasFeeFn({
@@ -974,11 +971,11 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
   readonly #handleEvmTransactionBatch = async (
     args: Omit<
       Parameters<typeof getAddTransactionBatchParams>[0],
-      'messagingSystem' | 'estimateGasFeeFn'
+      'messenger' | 'estimateGasFeeFn'
     >,
   ) => {
     const transactionParams = await getAddTransactionBatchParams({
-      messagingSystem: this.messagingSystem,
+      messenger: this.messenger,
       estimateGasFeeFn: this.#estimateGasFeeFn,
       ...args,
     });
@@ -1000,7 +997,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
     const { batchId } = await this.#addTransactionBatchFn(transactionParams);
 
     const { approvalMeta, tradeMeta } = findAndUpdateTransactionsInBatch({
-      messagingSystem: this.messagingSystem,
+      messenger: this.messenger,
       updateTransactionFn: this.#updateTransactionFn,
       batchId,
       txDataByType,
@@ -1028,7 +1025,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
     quoteResponse: QuoteResponse & Partial<QuoteMetadata>,
     isStxEnabledOnClient: boolean,
   ): Promise<TransactionMeta & Partial<SolanaTransactionMeta>> => {
-    this.messagingSystem.call('BridgeController:stopPollingForQuotes');
+    this.messenger.call('BridgeController:stopPollingForQuotes');
 
     const selectedAccount = this.#getMultichainSelectedAccount(accountAddress);
     if (!selectedAccount) {
@@ -1125,7 +1122,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
               await this.#handleEvmTransactionBatch({
                 isBridgeTx,
                 resetApproval: await getUSDTAllowanceResetTx(
-                  this.messagingSystem,
+                  this.messenger,
                   quoteResponse,
                 ),
                 approval: quoteResponse.approval,
@@ -1217,7 +1214,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
 
     // This will publish events for PERPS dropped tx failures as well
     if (!txMetaId) {
-      this.messagingSystem.call(
+      this.messenger.call(
         'BridgeController:trackUnifiedSwapBridgeEvent',
         eventName,
         baseProperties,
@@ -1228,7 +1225,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
     const historyItem: BridgeHistoryItem | undefined =
       this.state.txHistory[txMetaId];
     if (!historyItem) {
-      this.messagingSystem.call(
+      this.messenger.call(
         'BridgeController:trackUnifiedSwapBridgeEvent',
         eventName,
         eventProperties ?? {},
@@ -1245,7 +1242,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
         token_address_source,
         token_address_destination,
       } = requestParamProperties;
-      this.messagingSystem.call(
+      this.messenger.call(
         'BridgeController:trackUnifiedSwapBridgeEvent',
         eventName,
         {
@@ -1265,12 +1262,12 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
       return;
     }
 
-    const selectedAccount = this.messagingSystem.call(
+    const selectedAccount = this.messenger.call(
       'AccountsController:getAccountByAddress',
       historyItem.account,
     );
 
-    const { transactions } = this.messagingSystem.call(
+    const { transactions } = this.messenger.call(
       'TransactionController:getState',
     );
     const txMeta = transactions?.find(({ id }) => id === txMetaId);
@@ -1288,7 +1285,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
       ...getPriceImpactFromQuote(historyItem.quote),
     };
 
-    this.messagingSystem.call(
+    this.messenger.call(
       'BridgeController:trackUnifiedSwapBridgeEvent',
       eventName,
       requiredEventProperties,