+++ packages/bridge-status-controller/src/utils/transaction.ts
@@ -38,7 +38,7 @@ import type { Trade } from './trade-utils';
 export const generateActionId = () => (Date.now() + Math.random()).toString();
 
 export const getUSDTAllowanceResetTx = async (
-  messagingSystem: BridgeStatusControllerMessenger,
+  messenger: BridgeStatusControllerMessenger,
   quoteResponse: QuoteResponse & Partial<QuoteMetadata>,
 ) => {
   const hexChainId = formatChainIdToHex(quoteResponse.quote.srcChainId);
@@ -47,7 +47,7 @@ export const getUSDTAllowanceResetTx = async (
     isEthUsdt(hexChainId, quoteResponse.quote.srcAsset.address)
   ) {
     const allowance = new BigNumber(
-      await messagingSystem.call(
+      await messenger.call(
         'BridgeController:getBridgeERC20Allowance',
         quoteResponse.quote.srcAsset.address,
         hexChainId,
@@ -286,7 +286,7 @@ export const toBatchTxParams = (
 };
 
 export const getAddTransactionBatchParams = async ({
-  messagingSystem,
+  messenger,
   isBridgeTx,
   approval,
   resetApproval,
@@ -303,7 +303,7 @@ export const getAddTransactionBatchParams = async ({
   requireApproval = false,
   estimateGasFeeFn,
 }: {
-  messagingSystem: BridgeStatusControllerMessenger;
+  messenger: BridgeStatusControllerMessenger;
   isBridgeTx: boolean;
   trade: TxData;
   quoteResponse: Omit<QuoteResponse, 'approval' | 'trade'> &
@@ -314,7 +314,7 @@ export const getAddTransactionBatchParams = async ({
   requireApproval?: boolean;
 }) => {
   const isGasless = gasIncluded || gasIncluded7702;
-  const selectedAccount = messagingSystem.call(
+  const selectedAccount = messenger.call(
     'AccountsController:getAccountByAddress',
     trade.from,
   );
@@ -324,7 +324,7 @@ export const getAddTransactionBatchParams = async ({
     );
   }
   const hexChainId = formatChainIdToHex(trade.chainId);
-  const networkClientId = messagingSystem.call(
+  const networkClientId = messenger.call(
     'NetworkController:findNetworkClientIdByChainId',
     hexChainId,
   );
@@ -336,7 +336,7 @@ export const getAddTransactionBatchParams = async ({
   if (resetApproval) {
     const gasFees = await calculateGasFees(
       disable7702,
-      messagingSystem,
+      messenger,
       estimateGasFeeFn,
       resetApproval,
       networkClientId,
@@ -353,7 +353,7 @@ export const getAddTransactionBatchParams = async ({
   if (approval) {
     const gasFees = await calculateGasFees(
       disable7702,
-      messagingSystem,
+      messenger,
       estimateGasFeeFn,
       approval,
       networkClientId,
@@ -369,7 +369,7 @@ export const getAddTransactionBatchParams = async ({
   }
   const gasFees = await calculateGasFees(
     disable7702,
-    messagingSystem,
+    messenger,
     estimateGasFeeFn,
     trade,
     networkClientId,
@@ -400,19 +400,17 @@ export const getAddTransactionBatchParams = async ({
 };
 
 export const findAndUpdateTransactionsInBatch = ({
-  messagingSystem,
+  messenger,
   updateTransactionFn,
   batchId,
   txDataByType,
 }: {
-  messagingSystem: BridgeStatusControllerMessenger;
+  messenger: BridgeStatusControllerMessenger;
   updateTransactionFn: typeof TransactionController.prototype.updateTransaction;
   batchId: string;
   txDataByType: { [key in TransactionType]?: string };
 }) => {
-  const txs = messagingSystem.call(
-    'TransactionController:getState',
-  ).transactions;
+  const txs = messenger.call('TransactionController:getState').transactions;
   const txBatch: {
     approvalMeta?: TransactionMeta;
     tradeMeta?: TransactionMeta;